+++ constant/config.go
@@ -5,6 +5,7 @@ const (
 	DownloaderChapterNameTemplate = "downloader.chapter_name_template"
 	DownloaderAsync               = "downloader.async"
 	DownloaderCreateMangaDir      = "downloader.create_manga_dir"
+	DownloaderCreateVolumeDir     = "downloader.create_volume_dir"
 	DownloaderDefaultSource       = "downloader.default_source"
 	DownloaderStopOnError         = "downloader.stop_on_error"
 )
@@ -61,4 +62,5 @@ const (
 const (
 	LogsWrite = "logs.write"
 	LogsLevel = "logs.level"
+	LogsJson  = "logs.json"
 )