+++ luamodules/headless/element.go
@@ -33,7 +33,7 @@ func checkElement(L *lua.LState) *rod.Element {
 func elementInput(L *lua.LState) int {
 	el := checkElement(L)
 	value := L.ToString(2)
-	log.Info("inputting " + value)
+	log.Debug("inputting " + value)
 	el.MustInput(value)
 
 	return 0
@@ -41,7 +41,7 @@ func elementInput(L *lua.LState) int {
 
 func elementClick(L *lua.LState) int {
 	el := checkElement(L)
-	log.Info("clicking")
+	log.Debug("clicking")
 	el.MustClick()
 
 	return 0
@@ -49,7 +49,7 @@ func elementClick(L *lua.LState) int {
 
 func elementText(L *lua.LState) int {
 	el := checkElement(L)
-	log.Info("getting text")
+	log.Debug("getting text")
 	text := el.MustText()
 
 	L.Push(lua.LString(text))
@@ -59,7 +59,7 @@ func elementText(L *lua.LState) int {
 func elementAttribute(L *lua.LState) int {
 	el := checkElement(L)
 	name := L.ToString(2)
-	log.Info("getting attribute " + name)
+	log.Debug("getting attribute " + name)
 	value := el.MustAttribute(name)
 
 	L.Push(lua.LString(*value))
@@ -68,7 +68,7 @@ func elementAttribute(L *lua.LState) int {
 
 func elementHtml(L *lua.LState) int {
 	el := checkElement(L)
-	log.Info("getting html")
+	log.Debug("getting html")
 	html := el.MustHTML()
 
 	L.Push(lua.LString(html))
@@ -78,7 +78,7 @@ func elementHtml(L *lua.LState) int {
 func elementProperty(L *lua.LState) int {
 	el := checkElement(L)
 	name := L.ToString(2)
-	log.Info("getting property " + name)
+	log.Debug("getting property " + name)
 	value := el.MustProperty(name)
 
 	L.Push(lua.LString(value.Str()))