+++ updater/detect.go
@@ -21,6 +21,7 @@ const (
 	script
 )
 
+// detectInstallationMethod detects the installation method.
 func detectInstallationMethod() installationMethod {
 	for _, t := range []lo.Tuple2[installationMethod, func() bool]{
 		{scoop, isUnderScoop},
@@ -46,10 +47,12 @@ func detectInstallationMethod() installationMethod {
 	return unknown
 }
 
+// isUnderTermux returns true if mangal is running under termux.
 func isUnderTermux() (ok bool) {
 	return has("termux-setup-storage")
 }
 
+// isUnderHomebrew returns true if mangal is running under homebrew.
 func isUnderHomebrew() (ok bool) {
 	if !has("brew") {
 		return
@@ -63,6 +66,7 @@ func isUnderHomebrew() (ok bool) {
 	return strings.Contains(out, constant.Mangal)
 }
 
+// isUnderScoop returns true if mangal is running under scoop.
 func isUnderScoop() (ok bool) {
 	if !has("scoop") {
 		return false
@@ -76,11 +80,13 @@ func isUnderScoop() (ok bool) {
 	return strings.Contains(path, filepath.Join("scoop", "shims"))
 }
 
+// has returns true if the command exists.
 func has(command string) bool {
 	ok, err := exec.LookPath(command)
 	return err == nil && ok != ""
 }
 
+// execute the command and returns the output as string.
 func execute(command string, arguments ...string) (output string, err error) {
 	stdout := bytes.NewBufferString("")