+++ updater/update.go
@@ -3,12 +3,14 @@ package updater
 import (
 	"errors"
 	"fmt"
+	"github.com/metafates/mangal/constant"
 	"github.com/metafates/mangal/icon"
 	"net/http"
 	"os"
 	"os/exec"
 )
 
+// Update updates mangal to the latest version.
 func Update() (err error) {
 	method := detectInstallationMethod()
 
@@ -32,22 +34,25 @@ func Update() (err error) {
 	return
 }
 
+// updateHomebrew updates mangal using homebrew.
 func updateHomebrew() (err error) {
-	cmd := exec.Command("brew", "upgrade", "mangal")
+	cmd := exec.Command("brew", "upgrade", constant.Mangal)
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 	return cmd.Run()
 }
 
+// updateScoop updates mangal using scoop.
 func updateScoop() (err error) {
-	cmd := exec.Command("scoop", "update", "mangal")
+	cmd := exec.Command("scoop", "update", constant.Mangal)
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 	return cmd.Run()
 }
 
+// updateScript updates mangal using the script.
 func updateScript() (err error) {
-	res, err := http.Get("https://raw.githubusercontent.com/metafates/mangal/main/scripts/install")
+	res, err := http.Get(constant.InstallScriptURL)
 	if err != nil {
 		return err
 	}